moved shex into its own class - created once at startup
Touched a lot of things to get this to work.  Not really happy with solution but it works.

There is a TON of dead code in here.
