+++ minerva-server/src/main/java/org/geneontology/minerva/server/inferences/InferenceProviderCreatorImpl.java
@@ -27,16 +27,18 @@ public class InferenceProviderCreatorImpl implements InferenceProviderCreator {
 	private final Semaphore concurrentLock;
 	private final boolean useSLME;
 	private final String name;
+	private final ShexController shex;
 
-	InferenceProviderCreatorImpl(OWLReasonerFactory rf, int maxConcurrent, boolean useSLME, String name) {
+	InferenceProviderCreatorImpl(OWLReasonerFactory rf, int maxConcurrent, boolean useSLME, String name, ShexController shex) {
 		super();
 		this.rf = rf;
 		this.useSLME = useSLME;
 		this.name = name;
 		this.concurrentLock = new Semaphore(maxConcurrent);
+		this.shex = shex;
 	}
 
-	public static InferenceProviderCreator createElk(boolean useSLME) {
+	public static InferenceProviderCreator createElk(boolean useSLME, ShexController shex) {
 		String name;
 		if (useSLME) {
 			name = "ELK-SLME";
@@ -44,16 +46,16 @@ public class InferenceProviderCreatorImpl implements InferenceProviderCreator {
 		else {
 			name = "ELK";
 		}
-		return new InferenceProviderCreatorImpl(new ElkReasonerFactory(), 1, useSLME, name);
+		return new InferenceProviderCreatorImpl(new ElkReasonerFactory(), 1, useSLME, name, shex);
 	}
 	
-	public static InferenceProviderCreator createHermiT() {
+	public static InferenceProviderCreator createHermiT(ShexController shex) {
 		int maxConcurrent = Runtime.getRuntime().availableProcessors();
-		return createHermiT(maxConcurrent);
+		return createHermiT(maxConcurrent, shex);
 	}
 	
-	public static InferenceProviderCreator createHermiT(int maxConcurrent) {
-		return new InferenceProviderCreatorImpl(new org.semanticweb.HermiT.ReasonerFactory(), maxConcurrent, true, "Hermit-SLME");
+	public static InferenceProviderCreator createHermiT(int maxConcurrent, ShexController shex) {
+		return new InferenceProviderCreatorImpl(new org.semanticweb.HermiT.ReasonerFactory(), maxConcurrent, true, "Hermit-SLME", shex);
 	}
 
 	@Override
@@ -76,7 +78,7 @@ public class InferenceProviderCreatorImpl implements InferenceProviderCreator {
 						LOG.info("Done creating module: "+model.getModelId());
 					}
 					reasoner = rf.createReasoner(ont);
-					provider = MapInferenceProvider.create(reasoner, ont);
+					provider = MapInferenceProvider.create(reasoner, ont, shex);
 				}
 				finally {
 					concurrentLock.release();
